requirements.txt
appdirs==1.4.3
args==0.1.0
clint==0.5.1
httplib2==0.10.3
packaging==16.8
pyparsing==2.2.0
PySocks==1.6.7
pytz==2017.2
requests==2.13.0
simplejson==3.10.0
six==1.10.0
